Library 실습
1. We are going to implement a library system using classes of library and book. First, create “Book” class. The class should have following fields.
title as String
inPlace as boolean (a field to indicate whether a book is rented)
The class has following methods and constructor
Book constructor that takes book title as an input argument (argument name should be title) and set the title field
isInPlace() : a method to indicate if the book is borrowed
getTitle() : a method to return the title of a book
rent() : a method to mark the book as rented
handIn() : a method to mark the book as returned

The “Book” class should be tested with Homework4.java file provided in the slack #assignments channel

2. After completing problem 1, you have to implement a Library class that contains a set of Book objects created from class definition in problem 5 (arrays of book objects). In the library class, you have to add the following.
The Library class has a private field “name” as a String type to indicate the name of a library
The Library class has a public field “hour” as a String type to indicate the hours of the library. By default, set the hour as “9 AM to 5PM”. The hour variable is same for all libraries and it does not have to be an instance member
The Library class has a private field “books” as an array of Book objects where the initial size is 10. In the constructor, you have to initialize each element as null in order to check if an element is filled with a specific index
The Library class should have a constructor that takes the library name as input argument and set the name field.
You have to add a public method addBook that takes a Book object as input and add it to an element of books field. You need to iterate elements of books field to check if an element value is null. If an element is null, you can insert the input Book object to it.
You have to implement a public getName method to return the name of a library
You have to implement a public printAllBooks() method that prints out all information about books 
You have to implement a public printAllRentedBooks() method that prints out all information about only the rented books.
You have to implement a public rentBook method that takes a book name as an input argument. If the book is not rented (Book.inPlace is true), you have to return true and mark the book as rented (set Book.inPlace as false). If the book is already rented, you have to return false. 
You have to implement a public handInBook method that takes the title of a book as an input argument. If the book was rented (Book.inPlace is false), toggle the Book.inPlace variable to indicate the book is returned.


Cautious for the following
Some fields/methods should be static. 
Please be careful when comparing string object equality (remember reference type variable comparison)